Fix: 이벤트리스너가 바인딩되지 않는 이슈 해결 Resolve #58
